+++ es-ES/index.js
@@ -34,6 +34,7 @@ import roadmap from './roadmap'
 import events from './events'
 import fanManager from './fan_manager'
 import readingList from './reading_list'
+import personalNotes from './personalNotes'
 
 export default {
   locales: [
@@ -101,5 +102,6 @@ export default {
   ...roadmap,
   ...events,
   ...fanManager,
-  ...readingList
+  ...readingList,
+  ...personalNotes
 }