+++ cs-CZ/personalNotes.js
@@ -4,5 +4,13 @@ export default {
   'personalNotes.storyNotes': 'Poznámky k příběhu',
   'personalNotes.chapterNotes': 'Poznámky ke kapitole',
   'personalNotes.universeNotes': 'Poznámky k vesmíru',
-  'personalNotes.encyclopediaNotes': 'Encyclopedia Notes'
+  'personalNotes.encyclopediaNotes': 'Poznámky k encyklopédii',
+  'personalNotes.storyPersonalNotesHeading': 'Osobní poznámky k příběhům',
+  'personalNotes.encyclopediaPersonalNotesHeading': 'Osobní poznámky k Encyklopedii',
+  'personalNotes.searchPersonalNotesHeading': 'Hledat v osobních poznámkách',
+  'personalNotes.goToStory': 'Přejít na příběh',
+  'personalNotes.goToChapter': 'Přejít do kapitoly',
+  'personalNotes.goToEncyclopedia': 'Přejít do encykoplopédie',
+  'personalNotes.goToEncyclopediaEntry': 'Přejít na záznam v encyklopédii',
+  'personalNotes.closeDropdown': 'Zavřít seznam'
 };