+++ en/notifications.js
@@ -12,6 +12,10 @@ export default {
   'notifications.menu.workshop': 'Workshop',
   'notifications.menu.news': 'News & updates',
 
+  'notifications.menu.markUnread': 'Mark as unread',
+  'notifications.menu.markRead': 'Mark as read',
+  'notifications.menu.delete': 'Delete notification',
+
   // Notification messages
   'notification.workshopStoryStatus.toHiatus':
     'Your story {storyName} was automatically set to hiatus after 6 months of inactivity.',
@@ -22,6 +26,9 @@ export default {
     '{storyName} has been cancelled.',
   'notification.followingNewChapter.new':
     'New chapter in {storyName} has been published!',
-  'notification.workshopStory.newComment':
-    '{username} has commented on {storyName}'
+  'notification.workshopStory.newComment': 'has commented on {storyName}', // Before this text will be the username of the user
+  'notification.friendRequest': 'send you a friend request.', // Before this text will be the username of the user
+
+  // Strings for extra section
+  'notificationExtra.workshopStoryStatus.goToStory': 'Go to story'
 }