+++ es-ES/personalNotes.js
@@ -1,8 +1,16 @@
 export default {
-  'personalNotes.personalNotes': 'Personal Notes',
-  'personalNotes.chooseType': 'Choose the type of personal note',
-  'personalNotes.storyNotes': 'Story Notes',
-  'personalNotes.chapterNotes': 'Chapter Notes',
-  'personalNotes.universeNotes': 'Universe Notes',
-  'personalNotes.encyclopediaNotes': 'Encyclopedia Notes'
+  'personalNotes.personalNotes': 'Notas Personales',
+  'personalNotes.chooseType': 'Elige el tipo de notas personales',
+  'personalNotes.storyNotes': 'Notas de Historia',
+  'personalNotes.chapterNotes': 'Notas de Capítulo',
+  'personalNotes.universeNotes': 'Notas de Universo',
+  'personalNotes.encyclopediaNotes': 'Notas de Enciclopedia',
+  'personalNotes.storyPersonalNotesHeading': 'Notas Personales sobre las Historias',
+  'personalNotes.encyclopediaPersonalNotesHeading': 'Notas Personales sobre la Enciclopedia',
+  'personalNotes.searchPersonalNotesHeading': 'Buscar Notas Personales',
+  'personalNotes.goToStory': 'Ir a la historia',
+  'personalNotes.goToChapter': 'Ir a un capítulo',
+  'personalNotes.goToEncyclopedia': 'Ir a la Enciclopedia del Universo',
+  'personalNotes.goToEncyclopediaEntry': 'Ir a una Entrada de la Enciclopedia',
+  'personalNotes.closeDropdown': 'Cerrar el desplegable'
 };