+++ ja-JP/notifications.js
@@ -4,5 +4,24 @@ export default {
   'flashnews.startsAt': 'このメッセージの表示を開始：',
   'flashnews.endsAt': 'このメッセージの表示を停止：',
   'flashnews.newLanguage': '追加する新しい言語を選択する',
-  'flashnews.onlyDisplayOn': 'ニュースは、以下で選択された言語（選択されている場合）のみで表示され、他の言語ではデフォルト言語であっても表示されません。'
+  'flashnews.onlyDisplayOn': 'ニュースは、以下で選択された言語（選択されている場合）のみで表示され、他の言語ではデフォルト言語であっても表示されません。',
+  'notifications.menu.all': 'All notification',
+  'notifications.menu.following': 'Following',
+  'notifications.menu.workshop': 'Workshop',
+  'notifications.menu.news': 'News & updates',
+  'notifications.menu.markUnread': 'Mark as unread',
+  'notifications.menu.markRead': 'Mark as read',
+  'notifications.menu.delete': 'Delete notification',
+  // Notification messages
+  'notification.workshopStoryStatus.toHiatus': 'Your story {storyName} was automatically set to hiatus after 6 months of inactivity.',
+  'notification.workshopStoryStatus.toCancelled': 'Your story {storyName} was automatically set to cancelled after one year of inactivity.',
+  'notification.followingStoryStatus.toHiatus': '{storyName} is on hiatus.',
+  'notification.followingStoryStatus.toCancelled': '{storyName} has been cancelled.',
+  'notification.followingNewChapter.new': 'New chapter in {storyName} has been published!',
+  'notification.workshopStory.newComment': 'has commented on {storyName}',
+  // Before this text will be the username of the user
+  'notification.friendRequest': 'send you a friend request.',
+  // Before this text will be the username of the user
+  // Strings for extra section
+  'notificationExtra.workshopStoryStatus.goToStory': 'Go to story'
 };